requirements.txt
Django==4.0.6
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-summernote==0.8.20.0
django-taggit==3.1.0
djangorestframework==3.14.0
django-admin-interface==0.24.2
django-storages==1.13.2
boto3
mysqlclient==2.1.1
oauthlib==3.2.2
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
django-compressor
mysql-connector-python
beautifulsoup4
urllib3==1.26.14
azure-core==1.26.4
azure-storage-blob==12.16.0
azure-storage-queue==12.6.0
django-apptemplates
gunicorn>=20.0,<21.0
dj-database-url>=1.0,<2.0
whitenoise>=6.0,<7.0
psycopg2>=2.0,<3.0
djangorestframework